Ejercicio 5: El admin puede borrar acogimientos no aceptados

+++ app/Http/Controllers/API/AcogimientoController.php
@@ -84,6 +84,9 @@ class AcogimientoController extends Controller
      */
     public function destroy(Acogimiento $acogimiento)
     {
-        $acogimiento->delete();
+        $this->authorize('delete', $acogimiento);
+        if(!$acogimiento->aceptado) {
+            $acogimiento->delete();
+        }
     }
 }
